+++ packages/google-vertex/src/google-vertex-provider.test.ts
@@ -46,7 +46,7 @@ describe('google-vertex-provider', () => {
         provider: 'google.vertex.chat',
         baseURL:
           'https://test-location-aiplatform.googleapis.com/v1/projects/test-project/locations/test-location/publishers/google',
-        headers: expect.any(Object),
+        headers: expect.any(Function),
         generateId: expect.any(Function),
       }),
     );
@@ -71,7 +71,7 @@ describe('google-vertex-provider', () => {
       'test-embedding-model',
       expect.objectContaining({
         provider: 'google.vertex.embedding',
-        headers: expect.any(Object),
+        headers: expect.any(Function),
         baseURL:
           'https://test-location-aiplatform.googleapis.com/v1/projects/test-project/locations/test-location/publishers/google',
       }),
@@ -90,7 +90,7 @@ describe('google-vertex-provider', () => {
     expect(GoogleGenerativeAILanguageModel).toHaveBeenCalledWith(
       expect.anything(),
       expect.objectContaining({
-        headers: customHeaders,
+        headers: expect.any(Function),
       }),
     );
   });
@@ -155,7 +155,7 @@ describe('google-vertex-provider', () => {
         provider: 'google.vertex.image',
         baseURL:
           'https://test-location-aiplatform.googleapis.com/v1/projects/test-project/locations/test-location/publishers/google',
-        headers: expect.any(Object),
+        headers: expect.any(Function),
       }),
     );
   });
@@ -173,7 +173,7 @@ describe('google-vertex-provider', () => {
         provider: 'google.vertex.chat',
         baseURL:
           'https://aiplatform.googleapis.com/v1/projects/test-project/locations/global/publishers/google',
-        headers: expect.any(Object),
+        headers: expect.any(Function),
         generateId: expect.any(Function),
       }),
     );
@@ -190,7 +190,7 @@ describe('google-vertex-provider', () => {
       'test-embedding-model',
       expect.objectContaining({
         provider: 'google.vertex.embedding',
-        headers: expect.any(Object),
+        headers: expect.any(Function),
         baseURL:
           'https://aiplatform.googleapis.com/v1/projects/test-project/locations/global/publishers/google',
       }),
@@ -210,7 +210,7 @@ describe('google-vertex-provider', () => {
         provider: 'google.vertex.image',
         baseURL:
           'https://aiplatform.googleapis.com/v1/projects/test-project/locations/global/publishers/google',
-        headers: expect.any(Object),
+        headers: expect.any(Function),
       }),
     );
   });
@@ -240,7 +240,7 @@ describe('google-vertex-provider', () => {
         provider: 'google.vertex.chat',
         baseURL:
           'https://us-central1-aiplatform.googleapis.com/v1/projects/test-project/locations/us-central1/publishers/google',
-        headers: expect.any(Object),
+        headers: expect.any(Function),
         generateId: expect.any(Function),
       }),
     );